+++ src/BitcoinKernel/ScriptVerification/PrecomputedTransactionData.cs
@@ -1,4 +1,4 @@
-using BitcoinKernel.Primatives;
+using BitcoinKernel.Primitives;
 using BitcoinKernel.Exceptions;
 using BitcoinKernel.Interop;